[CATALOGS] Added the create function in each node

+++ application/modules/catalogs/controllers/IndexController.php
@@ -4,26 +4,35 @@ class Catalogs_IndexController extends Babel_Action
 {
     public function indexAction() {
         $model_catalogs = new Catalogs();
+
         $this->view->catalogs = $model_catalogs->selectRoots();
+
+        if ($this->auth <> null) {
+            $this->view->form = new Catalogs_Form_Create();
+        }
     }
 
     public function newAction() {
         $this->requireLogin();
 
         $request = $this->getRequest();
-        $form = new Catalogs_Form_Create();
-
         if ($request->isPost()) {
+            $form = new Catalogs_Form_Create();
             if ($form->isValid($request->getPost())) {
                 $model_catalogs = new Catalogs();
                 $catalog = $model_catalogs->createRow();
 
                 $catalog->label = $form->getSubForm('information')->getElement('label')->getValue();
+                $catalog->code = $form->getSubForm('information')->getElement('code')->getValue();
                 $catalog->description = $form->getSubForm('information')->getElement('description')->getValue();
                 $catalog->tsregister = time();
 
-                $iconv = new Yachay_Helpers_Iconv();
-                $catalog->url = $iconv->iconv($catalog->label);
+                // Parent comprobation
+                $id_parent = $request->getParam('catalog', 0);
+                $parent = $model_catalogs->findByIdent($id_parent);
+                if ($parent <> NULL) {
+                    $catalog->parent = $parent->ident;
+                }
 
                 $catalog->save();
                 $this->_helper->flashMessenger->addMessage('The catalog ' . $catalog->label . ' was created');
@@ -31,7 +40,7 @@ class Catalogs_IndexController extends Babel_Action
                 if ($form->getSubForm('photo')->getElement('photo')->receive()) {
                     $filename = $form->getSubForm('photo')->getElement('photo')->getFileName();
 
-                    if (file_exists($filename)) {
+                    if (!empty($filename) && file_exists($filename)) {
                         $thumbnail = new Yachay_Helpers_Thumbnail();
                         $thumbnail->thumbnail($filename, APPLICATION_PATH . '/../public/media/img/thumbnails/catalogs/' . $catalog->ident . '.jpg', 100, 100);
                         unlink($filename);
@@ -43,10 +52,17 @@ class Catalogs_IndexController extends Babel_Action
                     $this->_helper->flashMessenger->addMessage('The photo of catalog ' . $catalog->label . ' was updated successfully');
                 }
 
-                $this->_helper->redirector('index', 'index', 'catalogs');
+                if ($id_parent == 0) {
+                    $this->_helper->redirector('index', 'index', 'catalogs');
+                } else {
+                    $url = new Zend_Controller_Action_Helper_Url();
+                    $this->_redirect($url->url(array('catalog' => $id_parent), 'catalogs_catalog_view'));
+                }
+            } else {
+                $this->_helper->flashMessenger->addMessage('You must define a catalog name');
             }
         }
 
-        $this->view->form = $form;
+        $this->_helper->redirector('index', 'index', 'catalogs');
     }
 }
